[DEBT-043-D9] Split exa/client.py into focused modules
- Split 694-line exa/client.py into 6 focused mixin modules:
  - _http.py (228 LoC): HTTP request infrastructure + retry logic
  - _normalization.py (50 LoC): Content option normalization helpers
  - _search.py (262 LoC): search, search_and_contents, find_similar
  - _contents.py (100 LoC): get_contents endpoint
  - _answer.py (52 LoC): answer endpoint
  - _research.py (178 LoC): research task endpoints (create/get/list/wait)
- ExaClient now composes from all mixins (54 LoC)
- Updated test patches to use new module paths
- All 1052 tests pass, all quality gates pass
- Acceptance criteria: 1/1 complete
